requirements.txt
Flask
waitress
python-dotenv
schedule
psycopg2-binary
requests
urllib3
SQLAlchemy
aiohttp
aioschedule
sqlalchemy[asyncio]
asyncpg
openpyxl
python-dateutil
pillow